Extends makebuilder for cross-compiling

+++ src/makebuilder/handler/CppHandler.java
@@ -419,7 +419,7 @@ public class CppHandler implements SourceFileHandler {
             }
             be.target.addDependencies(dependencyBuffer);
         }
-        if (be.isUnitTest()) {
+        if (be.isUnitTest() && (!builder.isCrossCompiling())) {
             be.target.addCommand("echo ; echo \"===== Running unit test " + be.getTargetFilename() + " =====\"" , false);
             be.target.addCommand("" + be.getTarget() + " || ( rm " + be.getTarget() + " && false )" , false);
             be.target.addCommand("echo \"=====\" ; echo", false);
